+++ src/server.ts
@@ -39,7 +39,7 @@ import { log } from "./util/log.js";
 import { detectRolepodParent } from "./util/rolepodProtocol.js";
 
 export const SERVER_NAME = "rolepod-uiproof";
-export const SERVER_VERSION = "0.13.0";
+export const SERVER_VERSION = "0.13.1";
 
 /**
  * Extension Protocol version this build implements. Compared at server